chore: data types consolidation
Refs: XRDDEV-2929

+++ src/central-server/admin-service/infra-jpa/src/main/java/org/niis/xroad/cs/admin/jpa/repository/JpaSecurityServerClientRepository.java
@@ -45,7 +45,7 @@ import java.util.Optional;
 
 @Repository
 public interface JpaSecurityServerClientRepository<T extends SecurityServerClientEntity> extends
-        JpaRepository<T, Integer>,
+        JpaRepository<T, Long>,
         JpaSpecificationExecutor<T>,
         SecurityServerClientRepository<T> {
